+++ src/pdfmake.d.ts
@@ -8,10 +8,13 @@ declare module "pdfmake/build/pdfmake" {
   }
   const pdfMake: {
     vfs: Record<string, string>;
-    fonts?: unknown;
+    addVirtualFileSystem(vfs: Record<string, string>): void;
     createPdf(documentDefinition: TDocumentDefinitions): PdfDoc;
   };
   export default pdfMake;
 }
 
-declare module "pdfmake/build/vfs_fonts";
+declare module "pdfmake/build/vfs_fonts" {
+  const vfs: Record<string, string>;
+  export default vfs;
+}